requirements.txt
requests>=2.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.64.0
sqlalchemy>=2.0.0